+++ R/geodimension.R
@@ -720,3 +720,126 @@ get_level_data.geodimension <- function(gd,
   data
 }
 
+
+#' Get level layer
+#'
+#' Get a geographic layer associated with a level. We can select the geometry
+#' and, using boolean parameters, which attributes are included in the layer's
+#' table: only the attributes that make up the key, the subrogate key, inherited
+#' attributes.
+#'
+#' In case of inheriting attributes from other levels, in the table, these can
+#' have as a prefix the name of the level.
+#'
+#' @param gd A `geodimension` object.
+#' @param level_name A string.
+#' @param geometry A string.
+#' @param only_key A boolean.
+#' @param inherited A boolean.
+#' @param add_prefix A boolean.
+#'
+#' @return A `sf` object.
+#'
+#' @family information output functions
+#'
+#' @examples
+#'
+#' region <-
+#'   geolevel(name = "region",
+#'            layer = layer_us_region,
+#'            key = "geoid")
+#'
+#' division <-
+#'   geolevel(name = "division",
+#'            layer = layer_us_division,
+#'            key = "geoid")
+#'
+#' gd <-
+#'   geodimension(name = "gd_us",
+#'                level = region) |>
+#'   add_level(division)
+#'
+#' gd <- gd |>
+#'   relate_levels(lower_level_name = "division",
+#'                 lower_level_attributes = "REGION",
+#'                 upper_level_name = "region")
+#'
+#' ll <- gd |>
+#'   get_level_layer(level_name = "division",
+#'                   only_key = TRUE)
+#'
+#' @export
+get_level_layer <- function(gd,
+                            level_name,
+                            geometry,
+                            only_key,
+                            inherited,
+                            add_prefix) {
+  UseMethod("get_level_layer")
+}
+
+
+#' @rdname get_level_layer
+#' @export
+get_level_layer.geodimension <- function(gd,
+                                         level_name = NULL,
+                                         geometry = NULL,
+                                         only_key = FALSE,
+                                         inherited = FALSE,
+                                         add_prefix = TRUE) {
+  stopifnot("Missing level name." = !is.null(level_name))
+  level_name <-
+    validate_names(names(gd$geolevel), level_name, 'level')
+  if (is.null(geometry)) {
+    geometry <- names(gd$geolevel[[level_name]]$geometry)[1]
+  } else {
+    stopifnot("The selected geometry is not defined for the level." = geometry %in% names(gd$geolevel[[level_name]]$geometry))
+  }
+  layer <- gd$geolevel[[level_name]]$geometry[[geometry]]
+  if (!only_key) {
+    data <- gd |>
+      get_level_data(level_name = level_name, inherited = inherited, add_prefix = add_prefix)
+    layer <- data |>
+      dplyr::left_join(layer, by = gd$geolevel[[level_name]]$key) |>
+      sf::st_as_sf()
+
+  }
+  layer
+}
+
+
+
+#' Transform CRS
+#'
+#' Transform the CRS of all the layers included in the dimension to the one
+#' indicated.
+#'
+#' @param gd A `geodimension` object.
+#' @param crs A coordinate reference system: integer with the EPSG code, or
+#'   character with proj4string.
+#'
+#' @return A `geodimension`.
+#'
+#' @family configuration functions
+#'
+#' @export
+transform_crs <- function(gd,
+                          crs = NULL) {
+  UseMethod("transform_crs")
+}
+
+#' @rdname transform_crs
+#' @export
+transform_crs.geodimension <- function(gd,
+                                       crs = NULL) {
+  stopifnot("The crs is missing." = !is.null(crs))
+  for (layer in names(gd$geolevel)) {
+    for (geom in names(gd$geolevel[[layer]]$geometry)) {
+      gd$geolevel[[layer]]$geometry[[geom]] <-
+        gd$geolevel[[layer]]$geometry[[geom]] |>
+        sf::st_transform(crs = crs, use_gdal = FALSE)
+    }
+  }
+  gd
+}
+